+++ src/main/java/me/artograd/javamongoapi/controllers/UserController.java
@@ -1,7 +1,9 @@
 package me.artograd.javamongoapi.controllers;
 
+import me.artograd.javamongoapi.model.Tender;
 import me.artograd.javamongoapi.model.User;
 import me.artograd.javamongoapi.model.UserAttribute;
+import me.artograd.javamongoapi.model.system.UserTokenClaims;
 import me.artograd.javamongoapi.services.CognitoService;
 
 import java.util.List;
@@ -16,6 +18,9 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+
+import io.swagger.v3.oas.annotations.security.SecurityRequirement;
+import jakarta.servlet.http.HttpServletRequest;
 
 @RestController
 @RequestMapping(value = "/users")
@@ -35,7 +40,11 @@ public class UserController {
     }
     
     @DeleteMapping("/{username}")
-    public ResponseEntity<?> deleteUserById(@PathVariable String username) {
+    @SecurityRequirement(name = "bearerAuth")
+    public ResponseEntity<?> deleteUserById(@PathVariable String username, HttpServletRequest request) {
+    	
+    	if ( isDenied(username, request) ) { return new ResponseEntity<>(HttpStatus.FORBIDDEN);}
+    	
         boolean success = cognitoService.deleteUserByUsername(username);
         if (success) {
             return new ResponseEntity<>(HttpStatus.NO_CONTENT);
@@ -45,12 +54,31 @@ public class UserController {
     }
 
     @PutMapping("/{username}")
-    public ResponseEntity<?> updateUserAttributesById(@PathVariable String username, @RequestBody List<UserAttribute> attributes) {
+    @SecurityRequirement(name = "bearerAuth")
+    public ResponseEntity<?> updateUserAttributesById(@PathVariable String username, @RequestBody List<UserAttribute> attributes, HttpServletRequest request) {
+    	
+    	if ( isDenied(username, request) ) { return new ResponseEntity<>(HttpStatus.FORBIDDEN);}
+    	
         boolean success = cognitoService.updateUserAttributes(username, attributes);
         if (success) {
             return new ResponseEntity<>(HttpStatus.OK);
         } else {
             return new ResponseEntity<>("Error updating user attributes", HttpStatus.INTERNAL_SERVER_ERROR);
         }
+    }
+    
+    /**
+     * Check that operation is executed by profile owner 
+     * @param tenderId
+     * @param request
+     * @return
+     */
+    private boolean isDenied(String username, HttpServletRequest request) {
+    	UserTokenClaims claims = cognitoService.getUserTokenClaims(request);
+    	if ( claims.getUsername() == null || //username is not provided in token
+    		!claims.getUsername().equals( username )) {//request is made on behalf of different user
+    		return true;
+    	}
+    	return false;
     }
 }
